refactoring into search.go
